Protobuff_and_Grpc-master/RestVSGrpc/requirements.txt
grpcio==1.73.0
grpcio-tools==1.73.0
protobuf==6.31.1
setuptools==80.9.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.3.0
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3